Refactor: File Name Deduplication Session 2
- Renamed files inside anticheat, auction, daily, economy, and kit directories to remove redundant prefixes.
- Updated all related imports across the codebase.
- Updated plan.md to reflect Session 2 completion.
- Formatted, linted, and verified typescript compilation and tests.
