+++ src/core/events/playerSpawn.ts
@@ -8,7 +8,7 @@ import { sendMessage } from '@core/messaging.js';
 import { getOrCreatePlayer, updatePlayerData } from '@core/playerDataManager.js';
 import { getPlayerRank, updatePlayerNameTag } from '@core/rankManager.js';
 import { formatLocation, formatString } from '@core/utils.js';
-import { getKit, giveKitItems } from '@features/kit/kitsManager.js';
+import { getKit, giveKitItems } from '@features/kit/manager.js';
 import { checkAndKickBannedPlayer } from '@features/moderation/punishmentManager.js';
 
 export function handlePlayerJoin(player: mc.Player) {
