+++ src/core/ui/uiUtils.ts
@@ -19,7 +19,7 @@ import {
 import type { PlayerData } from '@core/playerDataManager.js';
 import ranksConfig from '@core/ranksConfig.default.js';
 import { showPanel } from '@core/uiManager.js';
-import { AnticheatConfig, getAnticheatConfig, saveAnticheatConfig } from '@features/anticheat/anticheatConfigLoader.js';
+import { AnticheatConfig, getAnticheatConfig, saveAnticheatConfig } from '@features/anticheat/configLoader.js';
 import { xrayConfig } from '@features/anticheat/xrayConfig.default.js';
 import { auctionHouseConfig } from '@features/auction/auctionHouseConfig.default.js';
 import { economyConfig } from '@features/economy/economyConfig.js';
