+++ src/features/anticheat/__tests__/MovementCheck.test.ts
@@ -16,7 +16,7 @@ vi.mock('@core/logger.js', () => ({
     errorLog: vi.fn()
 }));
 
-vi.mock('../anticheatConfigLoader.js', () => ({
+vi.mock('../configLoader.js', () => ({
     getAnticheatConfig: mockGetConfig
 }));
