+++ src/features/auction/ui/panel.ts
@@ -3,7 +3,7 @@ import { ActionFormData, ActionFormResponse, ModalFormData, ModalFormResponse } 
 
 import { getOrCreatePlayer } from '@core/playerDataManager.js';
 import { formatCurrency, formatTime, uiWait } from '@core/utils.js';
-import { AuctionListing, buyItem, cancelListing, claimMailbox, claimMailboxItem, getListings, getListingsCount, placeBid, SortOption } from '@features/auction/auctionManager.js';
+import { AuctionListing, buyItem, cancelListing, claimMailbox, claimMailboxItem, getListings, getListingsCount, placeBid, SortOption } from '@features/auction/manager.js';
 import { isDefined, isNonEmptyString } from '@lib/guards.js';
 
 const LISTINGS_PER_PAGE = 45;
