+++ src/features/daily/manager.ts
