+++ src/features/economy/commands/bounty.ts
@@ -5,7 +5,7 @@ import { sendMessage } from '@core/messaging.js';
 import { getOrCreatePlayer, getPlayerIdByName, getPlayerNameById, incrementPlayerBalance } from '@core/playerDataManager.js';
 import { resolveTarget } from '@core/utils.js';
 import * as bountyManager from '@features/economy/bountyManager.js';
-import { validateCurrencyAmount } from '@features/economy/economyUtils.js';
+import { validateCurrencyAmount } from '@features/economy/utils.js';
 import { isDefined, isNonEmptyString } from '@lib/guards.js';
 
 import { CommandExecutor, CustomCommand } from '@commands/commandManager.js';
