+++ src/features/economy/index.ts
@@ -1,5 +1,5 @@
 import { BountyPanelHandler } from '@features/economy/ui/bountyPanel.js';
-import { EconomyPanelHandler } from '@features/economy/ui/economyPanel.js';
+import { EconomyPanelHandler } from '@features/economy/ui/panel.js';
 import { panelRouter } from '@ui/PanelRouter.js';
 
 export function initialize() {
