+++ src/features/economy/ui/panel.ts
