+++ src/features/kit/itemsManager.ts
@@ -2,7 +2,7 @@ import * as mc from '@minecraft/server';
 
 import { getKitsConfig, saveKitsConfig } from '@core/configurations.js';
 import { debugLog, errorLog } from '@core/logger.js';
-import { Kit } from '@features/kit/kitAdminManager.js';
+import { Kit } from '@features/kit/adminManager.js';
 import { isDefined, isNonEmptyString } from '@lib/guards.js';
 
 const MAX_KIT_SLOTS = 36;
